[feature1283Strafing] + Updated WeaonAttackAction to support strafing
Added a state variable to keep track of whether the WAA is for a strafing
attack or not, also added an additional parameter to the static toHit method
signatures.  Added cases to toHitImpossible related to strafing, and updated
the to-hit modifiers for strafing.  One thing that remains to be added is
considering terrain modifiers at NOE (and possibly at other altitudes).


git-svn-id: svn://svn.code.sf.net/p/megamek/code/branches/feature1283Strafing@11204 c34074a7-3587-4b08-b2b8-7c33201a20e7