[feature1283Strafing] + Updated Mounted to support strafing
One issue is that, weapons get fired multiple times during a strafing run, and
currently this is disallowed with the Mounted.usedThisRound state variable and
the Mounted.isReady() method.  I added a boolean flag to Mounted.isReady, that
denotes whether the call is related to strafing or not.  If the isReady method
is called for a strafing run, the Mounted instance can still be ready if it
has been fired before.



git-svn-id: svn://svn.code.sf.net/p/megamek/code/branches/feature1283Strafing@11205 c34074a7-3587-4b08-b2b8-7c33201a20e7