+ Bug: Updated A* heuristic for aerodyne ASFs: paths shoulder be shorter now
Given the restrictiosn for aerodyne ASF movement on ground maps, there's no
good admissible heuristic (that I know of at least).  So, the A* comparator
defaults to ranking paths based on their MP used and arbitrating ties with
the path that has the most hexes moved.  This makes sense for ground units
where the more hexes moved could mean a better target movement modifier,
however it doesn't make sense for ASF, and can lead to unintuitive paths,
so I switched the fall-back to pick paths with fewer hexes moved, which
will lead to shorter paths for ASF, and is more likely to be inline with
what the user has in mind.



git-svn-id: svn://svn.code.sf.net/p/megamek/code/trunk@11609 c34074a7-3587-4b08-b2b8-7c33201a20e7