+ After clicking backwards button, path planner will only step back and turn
In the previous path planner, if the user was in walk/backwards modes, the PP
would only use those steps (along with turns).  In the current PP, in walk mode
it will only add forward steps, but in backwards, it will add backwards or
forwards, based on what gives the "shortest path."  This really isn't what the
user wants: when the user clicks backwards they generally intend for backwards
movement to be used.  The problem is that the destination doesn't include
facing, and often the user is moving backwards to get a particular facing and
the new PP ends up causing problems because often moving forward looks cheaper
without facing considered.



git-svn-id: svn://svn.code.sf.net/p/megamek/code/trunk@11610 c34074a7-3587-4b08-b2b8-7c33201a20e7